Collapse Portfolio file structure
