+++ archetypes/Portfolio/OverviewTable/styles.tsx
