+++ archetypes/Portfolio/TradeOverviewBanner/styles.tsx
